requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
frozenlist==1.8.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
gspread==6.1.0
idna==3.11
multidict==6.7.1
oauthlib==3.3.1
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.11.0
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
StrEnum==0.4.15
flask>=3.0.0
supabase>=2.4.0
twilio==9.10.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.22.0

